requirements.txt
Flask==3.0.0
requests==2.31.0
Werkzeug==3.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
gunicorn==21.2.0
Pillow==10.1.0
